feat: restructure home dashboard and sidebar
- Introduced a new home dashboard layout with a dedicated Dashboard.vue component.
- Migrated existing dashboard functionality from the previous Dashboard.vue to the new home structure.
- Added sidebar items management through a new sidebarItems.ts file for better organization.
- Updated TeamSwitcher component to conditionally render based on the current team.
- Removed unused code and components to streamline the application.
