+++ frontend/src/layouts/BaseLayout.vue
@@ -52,9 +52,8 @@
 </template>
 <script setup lang="ts">
 import { session } from "@/data/session";
-import { computed } from "vue";
 import { Popover, Sidebar, type SidebarProps } from "frappe-ui";
-import { EllipsisVertical, LayoutDashboard } from "lucide-vue-next";
+import { EllipsisVertical } from "lucide-vue-next";
 import type { PropType } from "vue";
 import Avatar from "@/components/ui/Avatar.vue";
 import TeamSwitcher from "@/components/team/TeamSwitcher.vue";
@@ -71,27 +70,5 @@ const props = defineProps({
         type: Array as PropType<SidebarSectionProps[]>,
         default: () => [],
     },
-});
-
-const sidebarSections = computed((): SidebarSectionProps[] => {
-    if (props.sidebarSections && props.sidebarSections.length > 0) {
-        return props.sidebarSections;
-    }
-
-    const _sections: SidebarSectionProps[] = [
-        {
-            label: "",
-            items: [
-                {
-                    label: "Dashboard",
-                    to: "/",
-                    isActive: true,
-                    icon: LayoutDashboard,
-                },
-            ],
-        },
-    ];
-
-    return _sections;
 });
 </script>
