+++ frontend/src/pages/team/ManageTeam.vue
@@ -1,0 +1,2 @@
+<script setup lang="ts"></script>
+<template></template>
